Properly start contact view intent
