+++ app/src/main/kotlin/com/simplemobiletools/smsmessenger/activities/ThreadActivity.kt
@@ -1286,31 +1286,6 @@ class ThreadActivity : SimpleActivity() {
         return participants
     }
 
-    fun startContactDetailsIntent(contact: SimpleContact) {
-        val simpleContacts = "com.simplemobiletools.contacts.pro"
-        val simpleContactsDebug = "com.simplemobiletools.contacts.pro.debug"
-        if (contact.rawId > 1000000 && contact.contactId > 1000000 && contact.rawId == contact.contactId &&
-            (isPackageInstalled(simpleContacts) || isPackageInstalled(simpleContactsDebug))
-        ) {
-            Intent().apply {
-                action = Intent.ACTION_VIEW
-                putExtra(CONTACT_ID, contact.rawId)
-                putExtra(IS_PRIVATE, true)
-                setPackage(if (isPackageInstalled(simpleContacts)) simpleContacts else simpleContactsDebug)
-                setDataAndType(ContactsContract.Contacts.CONTENT_LOOKUP_URI, "vnd.android.cursor.dir/person")
-                launchActivityIntent(this)
-            }
-        } else {
-            ensureBackgroundThread {
-                val lookupKey = SimpleContactsHelper(this).getContactLookupKey((contact).rawId.toString())
-                val publicUri = Uri.withAppendedPath(ContactsContract.Contacts.CONTENT_LOOKUP_URI, lookupKey)
-                runOnUiThread {
-                    launchViewContactIntent(publicUri)
-                }
-            }
-        }
-    }
-
     fun saveMMS(mimeType: String, path: String) {
         hideKeyboard()
         lastAttachmentUri = path
